+++ typescript/src/processors/readTypeScriptModules/index.ts
@@ -13,13 +13,15 @@ import { EnumExportDoc } from '../../api-doc-types/EnumExportDoc';
 import { ExportDoc } from '../../api-doc-types/ExportDoc';
 import { FunctionExportDoc } from '../../api-doc-types/FunctionExportDoc';
 import { InterfaceExportDoc } from '../../api-doc-types/InterfaceExportDoc';
+import { MemberDoc } from '../../api-doc-types/MemberDoc';
+import { MethodMemberDoc } from '../../api-doc-types/MethodMemberDoc';
 import { ModuleDoc } from '../../api-doc-types/ModuleDoc';
 import { TypeAliasExportDoc } from '../../api-doc-types/TypeAliasExportDoc';
 
 import { expandSourceFiles, SourcePattern } from './SourcePattern';
 
-export function readTypeScriptModules(tsParser: TsParser, modules: any, createDocMessage: any, log: any) {
-  return new ReadTypeScriptModules(tsParser, modules, createDocMessage, log);
+export function readTypeScriptModules(tsParser: TsParser, modules: any, namespacesToInclude: string[], createDocMessage: any, log: any) {
+  return new ReadTypeScriptModules(tsParser, modules, namespacesToInclude, createDocMessage, log);
 }
 
 export class ReadTypeScriptModules implements Processor {
@@ -49,6 +51,7 @@ export class ReadTypeScriptModules implements Processor {
     constructor(
       private tsParser: TsParser,
       private modules: any,
+      private namespacesToInclude: string[],
       private createDocMessage: any,
       private log: any) {}
 
@@ -94,32 +97,32 @@ export class ReadTypeScriptModules implements Processor {
 
         switch (getExportDocType(resolvedExport)) {
           case 'class':
-            const classDoc = new ClassExportDoc(moduleDoc, resolvedExport, this.basePath, this.hidePrivateMembers);
-            classDoc.members.forEach(doc => docs.push(doc));
-            classDoc.statics.forEach(doc => docs.push(doc));
-            if (classDoc.constructorDoc) docs.push(classDoc.constructorDoc);
+            const classDoc = new ClassExportDoc(moduleDoc, resolvedExport, this.basePath, this.hidePrivateMembers, this.namespacesToInclude);
+            this.addMemberDocs(docs, classDoc.members);
+            this.addMemberDocs(docs, classDoc.statics);
+            if (classDoc.constructorDoc) this.addMemberDocs(docs, [classDoc.constructorDoc]);
             this.addExportDoc(docs, moduleDoc, classDoc);
             break;
           case 'interface':
-            const interfaceDoc = new InterfaceExportDoc(moduleDoc, resolvedExport, this.basePath);
-            interfaceDoc.members.forEach(doc => docs.push(doc));
+            const interfaceDoc = new InterfaceExportDoc(moduleDoc, resolvedExport, this.basePath, this.namespacesToInclude);
+            this.addMemberDocs(docs, interfaceDoc.members);
             this.addExportDoc(docs, moduleDoc, interfaceDoc);
             break;
           case 'enum':
-            const enumDoc = new EnumExportDoc(moduleDoc, resolvedExport, this.basePath);
+            const enumDoc = new EnumExportDoc(moduleDoc, resolvedExport, this.basePath, this.namespacesToInclude);
             enumDoc.members.forEach(doc => docs.push(doc));
             this.addExportDoc(docs, moduleDoc, enumDoc);
             break;
           case 'const':
           case 'let':
           case 'var':
-            this.addExportDoc(docs, moduleDoc, new ConstExportDoc(moduleDoc, resolvedExport, this.basePath));
+            this.addExportDoc(docs, moduleDoc, new ConstExportDoc(moduleDoc, resolvedExport, this.basePath, this.namespacesToInclude));
             break;
           case 'type-alias':
-            this.addExportDoc(docs, moduleDoc, new TypeAliasExportDoc(moduleDoc, resolvedExport, this.basePath));
+            this.addExportDoc(docs, moduleDoc, new TypeAliasExportDoc(moduleDoc, resolvedExport, this.basePath, this.namespacesToInclude));
             break;
           case 'function':
-            const functionDoc = new FunctionExportDoc(moduleDoc, resolvedExport, this.basePath);
+            const functionDoc = new FunctionExportDoc(moduleDoc, resolvedExport, this.basePath, this.namespacesToInclude);
             this.addExportDoc(docs, moduleDoc, functionDoc);
             functionDoc.overloads.forEach(doc => docs.push(doc));
             break;
@@ -134,6 +137,13 @@ export class ReadTypeScriptModules implements Processor {
       this.log.debug('>>>> EXPORT: ' + exportDoc.name + ' (' + exportDoc.docType + ') from ' + moduleDoc.id);
       moduleDoc.exports.push(exportDoc);
       docs.push(exportDoc);
+    }
+
+    private addMemberDocs(docs: DocCollection, members: MemberDoc[]) {
+      members.forEach(member => {
+        docs.push(member);
+        if (member instanceof MethodMemberDoc) member.overloads.forEach(overloadDoc => docs.push(overloadDoc));
+      });
     }
   }